Move the middleware to db repo
