+++ package.json
@@ -26,7 +26,7 @@
     "@kadira/storybook-addon-links": "^1.0.0",
     "@kadira/storybook-addons": "^1.5.0",
     "@kadira/storybook-channel-pagebus": "^2.0.2",
-    "@kadira/storybook-database-local": "^1.0.3",
+    "@kadira/storybook-database-local": "^1.1.0",
     "@kadira/storybook-ui": "^3.3.2",
     "autoprefixer": "^6.3.7",
     "babel-core": "^6.11.4",
@@ -43,7 +43,6 @@
     "babel-preset-es2016": "^6.11.3",
     "babel-preset-react": "^6.11.1",
     "babel-runtime": "^6.9.2",
-    "body-parser": "^1.15.2",
     "case-sensitive-paths-webpack-plugin": "^1.1.2",
     "commander": "^2.9.0",
     "configstore": "^2.0.0",
@@ -55,7 +54,6 @@
     "json-stringify-safe": "^5.0.1",
     "json5": "^0.5.0",
     "lodash.pick": "^4.2.0",
-    "lowdb": "^0.13.1",
     "postcss-loader": "0.9.1",
     "qs": "^6.1.0",
     "react-modal": "^1.2.0",
