+++ src/server/middleware.js
@@ -2,11 +2,11 @@ import { Router } from 'express';
 import webpack from 'webpack';
 import webpackDevMiddleware from 'webpack-dev-middleware';
 import webpackHotMiddleware from 'webpack-hot-middleware';
-import baseConfig from '../config/webpack.config';
-import loadConfig from '../config';
-import getIndexHtml from '../index.html';
-import getIframeHtml from '../iframe.html';
-import { getHeadHtml } from '../utils';
+import baseConfig from './config/webpack.config';
+import loadConfig from './config';
+import getIndexHtml from './index.html';
+import getIframeHtml from './iframe.html';
+import { getHeadHtml } from './utils';
 
 export default function (configDir) {
   // Build the webpack configuration using the `baseConfig`
